Use long literals instead of .toLong

+++ core/shared/src/main/scala/org/threeten/bp/LocalDate.scala
@@ -672,7 +672,7 @@ final class LocalDate private (private val year: Int, monthOfYear: Int, dayOfMon
    *   the day-of-week, not null
    */
   def getDayOfWeek: DayOfWeek = {
-    val dow0: Int = Math.floorMod(toEpochDay + 3, 7.toLong).toInt
+    val dow0: Int = Math.floorMod(toEpochDay + 3, 7L).toInt
     DayOfWeek.of(dow0 + 1)
   }
 
@@ -990,9 +990,9 @@ final class LocalDate private (private val year: Int, monthOfYear: Int, dayOfMon
         case WEEKS     => plusWeeks(amountToAdd)
         case MONTHS    => plusMonths(amountToAdd)
         case YEARS     => plusYears(amountToAdd)
-        case DECADES   => plusYears(Math.multiplyExact(amountToAdd, 10.toLong))
-        case CENTURIES => plusYears(Math.multiplyExact(amountToAdd, 100.toLong))
-        case MILLENNIA => plusYears(Math.multiplyExact(amountToAdd, 1000.toLong))
+        case DECADES   => plusYears(Math.multiplyExact(amountToAdd, 10L))
+        case CENTURIES => plusYears(Math.multiplyExact(amountToAdd, 100L))
+        case MILLENNIA => plusYears(Math.multiplyExact(amountToAdd, 1000L))
         case ERAS      => `with`(ERA, Math.addExact(getLong(ERA), amountToAdd))
         case _         => throw new UnsupportedTemporalTypeException(s"Unsupported unit: $unit")
       }
@@ -1052,8 +1052,8 @@ final class LocalDate private (private val year: Int, monthOfYear: Int, dayOfMon
     else {
       val monthCount: Long = year * 12L + (month - 1)
       val calcMonths: Long = monthCount + monthsToAdd
-      val newYear: Int     = YEAR.checkValidIntValue(Math.floorDiv(calcMonths, 12.toLong))
-      val newMonth: Int    = Math.floorMod(calcMonths, 12.toLong).toInt + 1
+      val newYear: Int     = YEAR.checkValidIntValue(Math.floorDiv(calcMonths, 12L))
+      val newMonth: Int    = Math.floorMod(calcMonths, 12L).toInt + 1
       LocalDate.resolvePreviousValid(newYear, newMonth, day.toInt)
     }
 
@@ -1075,7 +1075,7 @@ final class LocalDate private (private val year: Int, monthOfYear: Int, dayOfMon
    * @throws DateTimeException
    *   if the result exceeds the supported date range
    */
-  def plusWeeks(weeksToAdd: Long): LocalDate = plusDays(Math.multiplyExact(weeksToAdd, 7.toLong))
+  def plusWeeks(weeksToAdd: Long): LocalDate = plusDays(Math.multiplyExact(weeksToAdd, 7L))
 
   /**
    * Returns a copy of this {@code LocalDate} with the specified number of days added.
